Fix data-template namespace issues
`data-template` is left in a rendered DOM element after being rendered in case
it needs to be refreshed later.  The only problem is that some other plug-ins
(namely Bootstrap's tooltip.js look for the data-template attribute when trying
to template out a related element (e.g. the tooltip).  I am not sure that this
commit is the best way to fix the problem as it is more of a bandage (simply
renaming the data attribute from `template` to `_renderingTemplate`.)  A better
solution might be to keep a private object of templates (keyed on the selector
or a unique id of the object) and not attaching it to the element's data object
directly.